Move delete_recipe to actions, rename client, and update tests
