+++ tests/routers/test_actions.py
@@ -24,6 +24,7 @@ from tests.constants import (
     RECIPES_SAVE_URL,
 )
 from tests.test_helpers import (
+    create_mock_api_response,
     extract_current_recipe_data_from_html,
     extract_full_edit_form_data,
 )
@@ -122,7 +123,7 @@ class TestSaveRecipeEndpoint:
             )
         )
         monkeypatch.setattr(
-            "meal_planner.routers.actions.internal_client.post", mock_post
+            "meal_planner.routers.actions.internal_api_client.post", mock_post
         )
 
         form_data = {
@@ -198,7 +199,7 @@ class TestSaveRecipeEndpoint:
         """Test error handling when the internal API call raises a generic exception."""
         mock_post = AsyncMock(side_effect=Exception("Unexpected network issue"))
         monkeypatch.setattr(
-            "meal_planner.routers.actions.internal_client.post", mock_post
+            "meal_planner.routers.actions.internal_api_client.post", mock_post
         )
 
         form_data = {
@@ -217,7 +218,7 @@ class TestSaveRecipeEndpoint:
         """Test error handling when the internal API call raises a RequestError."""
         mock_post = AsyncMock(side_effect=httpx.RequestError("Network issue"))
         monkeypatch.setattr(
-            "meal_planner.routers.actions.internal_client.post", mock_post
+            "meal_planner.routers.actions.internal_api_client.post", mock_post
         )
 
         form_data = {
@@ -250,7 +251,7 @@ class TestSaveRecipeEndpoint:
             )
         )
         monkeypatch.setattr(
-            "meal_planner.routers.actions.internal_client.post", mock_post
+            "meal_planner.routers.actions.internal_api_client.post", mock_post
         )
 
         form_data = {
@@ -281,7 +282,7 @@ class TestSaveRecipeEndpoint:
             )
         )
         monkeypatch.setattr(
-            "meal_planner.routers.actions.internal_client.post", mock_post
+            "meal_planner.routers.actions.internal_api_client.post", mock_post
         )
 
         form_data = {
@@ -313,7 +314,7 @@ class TestSaveRecipeEndpoint:
             )
         )
         monkeypatch.setattr(
-            "meal_planner.routers.actions.internal_client.post", mock_post
+            "meal_planner.routers.actions.internal_api_client.post", mock_post
         )
 
         form_data = {
@@ -1011,3 +1012,75 @@ def mock_recipe_data_fixture() -> RecipeBase:
         ingredients=["Mock Ing 1", "Mock Ing 2"],
         instructions=["Mock Inst 1.", "Mock Inst 2."],
     )
+
+
+@pytest.mark.anyio
+class TestDeleteRecipeEndpoint:
+    DELETE_PATH = "/recipes/delete"
+
+    @patch("meal_planner.routers.actions.internal_api_client", autospec=True)
+    async def test_delete_recipe_success(
+        self,
+        mock_api_client: AsyncMock,
+        client: AsyncClient,
+    ):
+        """Test successful recipe deletion."""
+        mock_api_client.delete.return_value = create_mock_api_response(status_code=204)
+
+        response = await client.post(self.DELETE_PATH, params={"id": 123})
+        assert response.status_code == 200
+        assert response.headers.get("HX-Trigger") == "recipeListChanged"
+        mock_api_client.delete.assert_called_once_with("/v0/recipes/123")
+
+    @patch("meal_planner.routers.actions.internal_api_client", autospec=True)
+    async def test_delete_recipe_not_found(
+        self,
+        mock_api_client: AsyncMock,
+        client: AsyncClient,
+    ):
+        """Test deletion of non-existent recipe."""
+        http_error = httpx.HTTPStatusError(
+            "Not Found",
+            request=httpx.Request("DELETE", "/v0/recipes/999"),
+            response=httpx.Response(404),
+        )
+        mock_api_client.delete.return_value = create_mock_api_response(
+            status_code=404, error_to_raise=http_error
+        )
+
+        response = await client.post(self.DELETE_PATH, params={"id": 999})
+        assert response.status_code == 404
+        mock_api_client.delete.assert_called_once_with("/v0/recipes/999")
+
+    @patch("meal_planner.routers.actions.internal_api_client", autospec=True)
+    async def test_delete_recipe_api_error(
+        self,
+        mock_api_client: AsyncMock,
+        client: AsyncClient,
+    ):
+        """Test API error during deletion."""
+        http_error = httpx.HTTPStatusError(
+            "Internal Server Error",
+            request=httpx.Request("DELETE", "/v0/recipes/123"),
+            response=httpx.Response(500),
+        )
+        mock_api_client.delete.return_value = create_mock_api_response(
+            status_code=500, error_to_raise=http_error
+        )
+
+        response = await client.post(self.DELETE_PATH, params={"id": 123})
+        assert response.status_code == 500
+        mock_api_client.delete.assert_called_once_with("/v0/recipes/123")
+
+    @patch("meal_planner.routers.actions.internal_api_client", autospec=True)
+    async def test_delete_recipe_generic_error(
+        self,
+        mock_api_client: AsyncMock,
+        client: AsyncClient,
+    ):
+        """Test generic error during deletion."""
+        mock_api_client.delete.side_effect = Exception("Generic API failure")
+
+        response = await client.post(self.DELETE_PATH, params={"id": 123})
+        assert response.status_code == 500
+        mock_api_client.delete.assert_called_once_with("/v0/recipes/123")
